requirements.txt
langchain
python-dotenv
dashscope
langchain-community
langchain_classic